# perf(es/minifier): Invoke DCE only before the first compression