--- a/crates/swc/tests/fixture/issues-5xxx/5865/output/index.js
+++ b/crates/swc/tests/fixture/issues-5xxx/5865/output/index.js
@@ -1,1 +1,1 @@
-var a;v=(a=r,b=>{let n=a.map(t=>{if(t)return t.foo});return n});
+var a;v=(a=r,b=>{let n=a.map(t=>{if(t){var e;return t.foo}});return n});