--- a/crates/swc/tests/fixture/sourcemap/004/output/index.js
+++ b/crates/swc/tests/fixture/sourcemap/004/output/index.js
@@ -36,6 +36,7 @@
             888,
             179
         ], function() {
+            var t;
             return n(n.s = 7139);
         }), _N_E = n.O();
     }