--- a/crates/swc/tests/tsc-references/Protected5.2.minified.js
+++ b/crates/swc/tests/tsc-references/Protected5.2.minified.js
@@ -1,1 +1,2 @@
 //// [Protected5.ts]
+import _class_call_check from "@swc/helpers/src/_class_call_check.mjs";