--- a/crates/swc/tests/tsc-references/accessorWithES5.2.minified.js
+++ b/crates/swc/tests/tsc-references/accessorWithES5.2.minified.js
@@ -1,1 +1,3 @@
 //// [accessorWithES5.ts]
+import _class_call_check from "@swc/helpers/src/_class_call_check.mjs";
+import _create_class from "@swc/helpers/src/_create_class.mjs";