--- a/crates/swc/tests/tsc-references/accessorsAreNotContextuallyTyped.2.minified.js
+++ b/crates/swc/tests/tsc-references/accessorsAreNotContextuallyTyped.2.minified.js
@@ -1,2 +1,5 @@
 //// [accessorsAreNotContextuallyTyped.ts]
-(void 0).x("");
+var c;
+import _class_call_check from "@swc/helpers/src/_class_call_check.mjs";
+import _create_class from "@swc/helpers/src/_create_class.mjs";
+c.x("");