--- a/crates/swc/tests/tsc-references/arrowFunctionExpressions.2.minified.js
+++ b/crates/swc/tests/tsc-references/arrowFunctionExpressions.2.minified.js
@@ -1,2 +1,3 @@
 //// [arrowFunctionExpressions.ts]
+import _class_call_check from "@swc/helpers/src/_class_call_check.mjs";
 0..toExponential();