--- a/crates/swc/tests/tsc-references/assignmentCompatWithConstructSignatures4.2.minified.js
+++ b/crates/swc/tests/tsc-references/assignmentCompatWithConstructSignatures4.2.minified.js
@@ -1,5 +1,8 @@
 //// [assignmentCompatWithConstructSignatures4.ts]
 var Errors;
+import _class_call_check from "@swc/helpers/src/_class_call_check.mjs";
+import _inherits from "@swc/helpers/src/_inherits.mjs";
+import _create_super from "@swc/helpers/src/_create_super.mjs";
 !function(Errors) {
     var WithNonGenericSignaturesInBaseType, WithGenericSignaturesInBaseType;
     WithNonGenericSignaturesInBaseType || (WithNonGenericSignaturesInBaseType = {}), WithGenericSignaturesInBaseType || (WithGenericSignaturesInBaseType = {});