--- a/crates/swc/tests/tsc-references/assignmentCompatWithConstructSignatures6.2.minified.js
+++ b/crates/swc/tests/tsc-references/assignmentCompatWithConstructSignatures6.2.minified.js
@@ -1,3 +1,6 @@
 //// [assignmentCompatWithConstructSignatures6.ts]
 var x, b, b2, b3, b4, b5, b11, b16;
+import _class_call_check from "@swc/helpers/src/_class_call_check.mjs";
+import _inherits from "@swc/helpers/src/_inherits.mjs";
+import _create_super from "@swc/helpers/src/_create_super.mjs";
 x.a = b, b = x.a, x.a2 = b2, b2 = x.a2, x.a3 = b3, b3 = x.a3, x.a4 = b4, b4 = x.a4, x.a5 = b5, b5 = x.a5, x.a11 = b11, b11 = x.a11, x.a16 = b16, b16 = x.a16;