--- a/crates/swc/tests/tsc-references/asyncArrowFunction1_es6.2.minified.js
+++ b/crates/swc/tests/tsc-references/asyncArrowFunction1_es6.2.minified.js
@@ -1,1 +1,2 @@
 //// [asyncArrowFunction1_es6.ts]
+import _async_to_generator from "@swc/helpers/src/_async_to_generator.mjs";