--- a/crates/swc/tests/tsc-references/asyncArrowFunctionCapturesThis_es5.2.minified.js
+++ b/crates/swc/tests/tsc-references/asyncArrowFunctionCapturesThis_es5.2.minified.js
@@ -1,1 +1,4 @@
 //// [asyncArrowFunctionCapturesThis_es5.ts]
+import _async_to_generator from "@swc/helpers/src/_async_to_generator.mjs";
+import _class_call_check from "@swc/helpers/src/_class_call_check.mjs";
+import _ts_generator from "@swc/helpers/src/_ts_generator.mjs";