--- a/crates/swc/tests/tsc-references/asyncAwaitIsolatedModules_es6.2.minified.js
+++ b/crates/swc/tests/tsc-references/asyncAwaitIsolatedModules_es6.2.minified.js
@@ -1,4 +1,4 @@
 //// [asyncAwaitIsolatedModules_es6.ts]
 var M;
+import _async_to_generator from "@swc/helpers/src/_async_to_generator.mjs";
 M || (M = {});
-export { };