--- a/crates/swc/tests/tsc-references/asyncFunctionDeclarationCapturesArguments_es5.2.minified.js
+++ b/crates/swc/tests/tsc-references/asyncFunctionDeclarationCapturesArguments_es5.2.minified.js
@@ -1,1 +1,2 @@
 //// [asyncFunctionDeclarationCapturesArguments_es5.ts]
+import _class_call_check from "@swc/helpers/src/_class_call_check.mjs";