--- a/crates/swc/tests/tsc-references/asyncMethodWithSuper_es5.2.minified.js
+++ b/crates/swc/tests/tsc-references/asyncMethodWithSuper_es5.2.minified.js
@@ -1,1 +1,9 @@
 //// [asyncMethodWithSuper_es5.ts]
+import _async_to_generator from "@swc/helpers/src/_async_to_generator.mjs";
+import _class_call_check from "@swc/helpers/src/_class_call_check.mjs";
+import _get from "@swc/helpers/src/_get.mjs";
+import _get_prototype_of from "@swc/helpers/src/_get_prototype_of.mjs";
+import _inherits from "@swc/helpers/src/_inherits.mjs";
+import _set from "@swc/helpers/src/_set.mjs";
+import _create_super from "@swc/helpers/src/_create_super.mjs";
+import _ts_generator from "@swc/helpers/src/_ts_generator.mjs";