--- a/crates/swc/tests/tsc-references/asyncUnParenthesizedArrowFunction_es5.2.minified.js
+++ b/crates/swc/tests/tsc-references/asyncUnParenthesizedArrowFunction_es5.2.minified.js
@@ -1,1 +1,3 @@
 //// [asyncUnParenthesizedArrowFunction_es5.ts]
+import _async_to_generator from "@swc/helpers/src/_async_to_generator.mjs";
+import _ts_generator from "@swc/helpers/src/_ts_generator.mjs";