--- a/crates/swc/tests/tsc-references/bestCommonTypeOfTuple2.2.minified.js
+++ b/crates/swc/tests/tsc-references/bestCommonTypeOfTuple2.2.minified.js
@@ -1,3 +1,6 @@
 //// [bestCommonTypeOfTuple2.ts]
-var t2, t3, t4, t5;
-(void 0)[4], t2[4], t3[4], t4[2], t5[2];
+var t1, t2, t3, t4, t5;
+import _class_call_check from "@swc/helpers/src/_class_call_check.mjs";
+import _inherits from "@swc/helpers/src/_inherits.mjs";
+import _create_super from "@swc/helpers/src/_create_super.mjs";
+t1[4], t2[4], t3[4], t4[2], t5[2];