--- a/crates/swc/tests/tsc-references/callSignaturesWithOptionalParameters.2.minified.js
+++ b/crates/swc/tests/tsc-references/callSignaturesWithOptionalParameters.2.minified.js
@@ -1,4 +1,5 @@
 //// [callSignaturesWithOptionalParameters.ts]
+import _class_call_check from "@swc/helpers/src/_class_call_check.mjs";
 function foo(x) {}
 var c, i, a, f = function(x) {}, f2 = function(x, y) {};
 foo(1), foo(), f(1), f(), f2(1), f2(1, 2), c.foo(), c.foo(1), i(), i(1), i.foo(1), i.foo(1, 2), a(), a(1), a.foo(), a.foo(1);