--- a/crates/swc/tests/tsc-references/callWithSpread.2.minified.js
+++ b/crates/swc/tests/tsc-references/callWithSpread.2.minified.js
@@ -1,6 +1,11 @@
 //// [callWithSpread.ts]
 var _instance, _instance1, _instance2, _instance3, _instance4, _instance5, _instance6, a, obj, xa;
+import _class_call_check from "@swc/helpers/src/_class_call_check.mjs";
+import _get from "@swc/helpers/src/_get.mjs";
+import _get_prototype_of from "@swc/helpers/src/_get_prototype_of.mjs";
+import _inherits from "@swc/helpers/src/_inherits.mjs";
 import _to_consumable_array from "@swc/helpers/src/_to_consumable_array.mjs";
+import _create_super from "@swc/helpers/src/_create_super.mjs";
 function foo(x, y) {
     for(var _len = arguments.length, z = Array(_len > 2 ? _len - 2 : 0), _key = 2; _key < _len; _key++)z[_key - 2] = arguments[_key];
 }