--- a/crates/swc/tests/tsc-references/castingTuple.2.minified.js
+++ b/crates/swc/tests/tsc-references/castingTuple.2.minified.js
@@ -1,5 +1,7 @@
 //// [castingTuple.ts]
 import _class_call_check from "@swc/helpers/src/_class_call_check.mjs";
+import _inherits from "@swc/helpers/src/_inherits.mjs";
+import _create_super from "@swc/helpers/src/_create_super.mjs";
 var E1, E2, C = function C() {
     "use strict";
     _class_call_check(this, C);