--- a/crates/swc/tests/tsc-references/checkJsxGenericTagHasCorrectInferences.2.minified.js
+++ b/crates/swc/tests/tsc-references/checkJsxGenericTagHasCorrectInferences.2.minified.js
@@ -1,1 +1,2 @@
 //// [file.tsx]
+import * as React from "react";