--- a/crates/swc/tests/tsc-references/checkJsxSubtleSkipContextSensitiveBug.2.minified.js
+++ b/crates/swc/tests/tsc-references/checkJsxSubtleSkipContextSensitiveBug.2.minified.js
@@ -1,3 +1,8 @@
 //// [checkJsxSubtleSkipContextSensitiveBug.tsx]
+import _async_to_generator from "@swc/helpers/src/_async_to_generator.mjs";
+import _class_call_check from "@swc/helpers/src/_class_call_check.mjs";
+import _inherits from "@swc/helpers/src/_inherits.mjs";
+import _create_super from "@swc/helpers/src/_create_super.mjs";
+import _ts_generator from "@swc/helpers/src/_ts_generator.mjs";
 import * as React from "react";
 React.Component;