--- a/crates/swc/tests/tsc-references/classAbstractAssignabilityConstructorFunction.2.minified.js
+++ b/crates/swc/tests/tsc-references/classAbstractAssignabilityConstructorFunction.2.minified.js
@@ -1,1 +1,2 @@
 //// [classAbstractAssignabilityConstructorFunction.ts]
+import _class_call_check from "@swc/helpers/src/_class_call_check.mjs";