--- a/crates/swc/tests/tsc-references/classAbstractOverloads.2.minified.js
+++ b/crates/swc/tests/tsc-references/classAbstractOverloads.2.minified.js
@@ -1,1 +1,2 @@
 //// [classAbstractOverloads.ts]
+import _class_call_check from "@swc/helpers/src/_class_call_check.mjs";