--- a/crates/swc/tests/tsc-references/classAbstractOverrideWithAbstract.2.minified.js
+++ b/crates/swc/tests/tsc-references/classAbstractOverrideWithAbstract.2.minified.js
@@ -1,1 +1,4 @@
 //// [classAbstractOverrideWithAbstract.ts]
+import _class_call_check from "@swc/helpers/src/_class_call_check.mjs";
+import _inherits from "@swc/helpers/src/_inherits.mjs";
+import _create_super from "@swc/helpers/src/_create_super.mjs";