--- a/crates/swc/tests/tsc-references/classAbstractSuperCalls.2.minified.js
+++ b/crates/swc/tests/tsc-references/classAbstractSuperCalls.2.minified.js
@@ -1,1 +1,6 @@
 //// [classAbstractSuperCalls.ts]
+import _class_call_check from "@swc/helpers/src/_class_call_check.mjs";
+import _get from "@swc/helpers/src/_get.mjs";
+import _get_prototype_of from "@swc/helpers/src/_get_prototype_of.mjs";
+import _inherits from "@swc/helpers/src/_inherits.mjs";
+import _create_super from "@swc/helpers/src/_create_super.mjs";