--- a/crates/swc/tests/tsc-references/classConstructorAccessibility5.2.minified.js
+++ b/crates/swc/tests/tsc-references/classConstructorAccessibility5.2.minified.js
@@ -1,1 +1,4 @@
 //// [classConstructorAccessibility5.ts]
+import _class_call_check from "@swc/helpers/src/_class_call_check.mjs";
+import _inherits from "@swc/helpers/src/_inherits.mjs";
+import _create_super from "@swc/helpers/src/_create_super.mjs";