--- a/crates/swc/tests/tsc-references/classConstructorParametersAccessibility2.2.minified.js
+++ b/crates/swc/tests/tsc-references/classConstructorParametersAccessibility2.2.minified.js
@@ -1,3 +1,6 @@
 //// [classConstructorParametersAccessibility2.ts]
-var c2;
-(void 0).x, c2.p, (void 0).p;
+var c1, c2;
+import _class_call_check from "@swc/helpers/src/_class_call_check.mjs";
+import _inherits from "@swc/helpers/src/_inherits.mjs";
+import _create_super from "@swc/helpers/src/_create_super.mjs";
+c1.x, c2.p, (void 0).p;