--- a/crates/swc/tests/tsc-references/classExtendingBuiltinType.2.minified.js
+++ b/crates/swc/tests/tsc-references/classExtendingBuiltinType.2.minified.js
@@ -1,3 +1,6 @@
 //// [classExtendingBuiltinType.ts]
+import _class_call_check from "@swc/helpers/src/_class_call_check.mjs";
+import _inherits from "@swc/helpers/src/_inherits.mjs";
 import _wrap_native_super from "@swc/helpers/src/_wrap_native_super.mjs";
+import _create_super from "@swc/helpers/src/_create_super.mjs";
 _wrap_native_super(Object), _wrap_native_super(Function), _wrap_native_super(String), _wrap_native_super(Boolean), _wrap_native_super(Number), _wrap_native_super(Date), _wrap_native_super(RegExp), _wrap_native_super(Error), _wrap_native_super(Array), _wrap_native_super(Array);