--- a/crates/swc/tests/tsc-references/classStaticBlock11(target=es2015).2.minified.js
+++ b/crates/swc/tests/tsc-references/classStaticBlock11(target=es2015).2.minified.js
@@ -1,1 +1,2 @@
 //// [classStaticBlock11.ts]
+import _class_private_field_get from "@swc/helpers/src/_class_private_field_get.mjs";