--- a/crates/swc_ecma_minifier/src/compress/mod.rs
+++ b/crates/swc_ecma_minifier/src/compress/mod.rs
@@ -269,50 +269,6 @@ where
             n.visit_with(&mut AssertValid);
         }
 
-        if self.options.unused {
-            let _timer = timer!("remove dead code");
-
-            let mut visitor = swc_ecma_transforms_optimization::simplify::dce::dce(
-                swc_ecma_transforms_optimization::simplify::dce::Config {
-                    module_mark: None,
-                    top_level: self.options.top_level(),
-                    top_retain: self.options.top_retain.clone(),
-                },
-                self.marks.unresolved_mark,
-            );
-
-            loop {
-                #[cfg(feature = "debug")]
-                let start = n.dump();
-
-                n.apply(&mut visitor);
-
-                self.changed |= visitor.changed();
-
-                #[cfg(feature = "debug")]
-                if visitor.changed() {
-                    let src = n.dump();
-                    debug!(
-                        "===== Before DCE =====\n{}\n===== After DCE =====\n{}",
-                        start, src
-                    );
-                }
-
-                if visitor.changed() {
-                    self.changed = true;
-                } else {
-                    break;
-                }
-
-                visitor.reset();
-            }
-        }
-
-        #[cfg(debug_assertions)]
-        {
-            n.visit_with(&mut AssertValid);
-        }
-
         {
             let _timer = timer!("apply full optimizer");