… crates/swc/tests/fixture/sourcemap/004/output/index.map
… crates/swc/tests/tsc-references/classExtendingClassLikeType.2.minified.js
… crates/swc/tests/tsc-references/classExtendingNonConstructor.2.minified.js
… crates/swc/tests/tsc-references/classExtendingNull.2.minified.js
… crates/swc/tests/tsc-references/classExtendsItself.2.minified.js
… crates/swc/tests/tsc-references/classExtendsItselfIndirectly.2.minified.js
… crates/swc/tests/tsc-references/classExtendsItselfIndirectly3.2.minified.js
… crates/swc/tests/tsc-references/classExtendsShadowedConstructorFunction.2.minified.js
… crates/swc/tests/tsc-references/classExtendsValidConstructorFunction.2.minified.js
… crates/swc/tests/tsc-references/classIsSubtypeOfBaseType.2.minified.js
… crates/swc/tests/tsc-references/classWithOnlyPublicMembersEquivalentToInterface.2.minified.js
… crates/swc/tests/tsc-references/classWithOnlyPublicMembersEquivalentToInterface2.2.minified.js
… crates/swc/tests/tsc-references/classWithOptionalParameter.2.minified.js
… crates/swc/tests/tsc-references/classWithProtectedProperty.2.minified.js
… crates/swc/tests/tsc-references/classWithStaticFieldInParameterInitializer(target=es5).2.minified.js
… crates/swc/tests/tsc-references/comparisonOperatorWithIdenticalObjects.2.minified.js
… crates/swc/tests/tsc-references/comparisonOperatorWithNoRelationshipObjectsOnCallSignature.2.minified.js
… crates/swc/tests/tsc-references/comparisonOperatorWithNoRelationshipObjectsOnConstructorSignature.2.minified.js
… crates/swc/tests/tsc-references/comparisonOperatorWithNoRelationshipObjectsOnIndexSignature.2.minified.js
… crates/swc/tests/tsc-references/comparisonOperatorWithNoRelationshipObjectsOnInstantiatedCallSignature.2.minified.js
… crates/swc/tests/tsc-references/comparisonOperatorWithNoRelationshipObjectsOnInstantiatedConstructorSignature.2.minified.js
… crates/swc/tests/tsc-references/comparisonOperatorWithSubtypeObjectOnCallSignature.2.minified.js
… crates/swc/tests/tsc-references/comparisonOperatorWithSubtypeObjectOnConstructorSignature.2.minified.js
… crates/swc/tests/tsc-references/comparisonOperatorWithSubtypeObjectOnIndexSignature.2.minified.js
… crates/swc/tests/tsc-references/comparisonOperatorWithSubtypeObjectOnInstantiatedCallSignature.2.minified.js
… crates/swc/tests/tsc-references/comparisonOperatorWithSubtypeObjectOnInstantiatedConstructorSignature.2.minified.js
… crates/swc/tests/tsc-references/comparisonOperatorWithSubtypeObjectOnProperty.2.minified.js
… crates/swc/tests/tsc-references/computedPropertyNames12_ES5.2.minified.js
… crates/swc/tests/tsc-references/computedPropertyNames13_ES5.2.minified.js
… crates/swc/tests/tsc-references/computedPropertyNames14_ES5.2.minified.js